+++ FinancialSummaryApi/V1/Boundary/Response/StatementListResponse.cs
@@ -4,7 +4,7 @@ namespace FinancialSummaryApi.V1.Boundary.Response
 {
     public class StatementListResponse
     {
-        public long Total { get; set; }
+        public int Total { get; set; }
         public List<StatementResponse> Statements { get; set; }
 
         public StatementListResponse()
